refactor: Enhance event handling and status management; update data structures and utility functions
